+++ src/components/postCard/postCard.tsx
@@ -1,6 +1,7 @@
 import { component$ } from '@builder.io/qwik';
 import { Link } from '@builder.io/qwik-city';
 import { HiArrowTopRightOnSquareOutline } from '@qwikest/icons/heroicons';
+import { textSm } from '~/style/style';
 import { css } from '~/styled-system/css';
 import { convertDateDisplay } from '~/utils/convertDateDisplay';
 
@@ -26,13 +27,13 @@ export default component$(({ title, href, createdAt, tags }: Props) => {
           </h2>
         </header>
         <div class={timeContainer}>
-          <time dateTime={createdAt} class={time}>
+          <time dateTime={createdAt} class={textSm}>
             {dateDisplay}
           </time>
           <div class={cardActions}>
             {tags.map((tag) => (
               <div key={tag} class={css({ color: '#f0abfc' })}>
-                <span class={hashTag}>#</span>
+                <span class={[hashTag, textSm]}>#</span>
                 <span>{tag}</span>
               </div>
             ))}
@@ -98,11 +99,6 @@ const timeContainer = css({
   justifyContent: 'end',
 });
 
-const time = css({
-  fontSize: '0.875rem',
-  lineHeight: '1.25rem',
-});
-
 const cardActions = css({
   display: 'flex',
   flexWrap: 'wrap',
@@ -112,7 +108,5 @@ const cardActions = css({
 });
 
 const hashTag = css({
-  fontSize: '0.875rem',
-  lineHeight: '1.25rem',
   mr: '1px',
 });
