+++ src/global.css
@@ -3,8 +3,3 @@ base,
 tokens,
 recipes,
 utilities;
-
-body {
-    font-family: 'Overpass', 'Noto Sans JP', sans-serif;
-    color: #bdc6e9;
-}
