+++ src/routes/blog/[postId]/index.tsx
@@ -2,7 +2,9 @@ import { component$ } from '@builder.io/qwik';
 import type { DocumentHead, StaticGenerateHandler } from '@builder.io/qwik-city';
 import { routeLoader$ } from '@builder.io/qwik-city';
 import { getPostDetail, getPostList } from '~/api/client';
+import PostContainer from '~/components/postContainer/postContainer';
 import { OG_IMAGE } from '~/const/seo';
+import { textSm } from '~/style/style';
 import { css } from '~/styled-system/css';
 import { divider } from '~/styled-system/patterns';
 import type { MyPost } from '~/types';
@@ -10,7 +12,7 @@ import { convertDateDisplay } from '~/utils/convertDateDisplay';
 
 export const usePostLoader = routeLoader$(async ({ params, status }) => {
   if (!params.postId) {
-    throw new Error('postId is required');
+    status(404);
   }
 
   try {
@@ -35,7 +37,7 @@ export default component$(() => {
       <header class={header}>
         <h1 class={title}>{post.value.title}</h1>
         <div class={infoContainer}>
-          <p class={infoText}>Published</p>
+          <p class={[infoText, textSm]}>Published</p>
           <time dateTime={post.value.publishedAt}>{dateDisplay}</time>
         </div>
       </header>
@@ -48,8 +50,8 @@ export default component$(() => {
           color: '#2A2F40',
         })}
       ></div>
-      <div class={postContainer}>
-        <div class={postBody} dangerouslySetInnerHTML={post.value.content}></div>
+      <div class={postWrapper}>
+        <PostContainer postContent={post.value.content} />
       </div>
     </article>
   );
@@ -146,132 +148,14 @@ const infoContainer = css({
 });
 
 const infoText = css({
-  fontSize: '0.875rem',
-  lineHeight: '1.25rem',
   fontWeight: '700',
   opacity: '0.6',
 });
 
-const postContainer = css({
+const postWrapper = css({
   mt: '5rem',
   w: '100%',
   fontSize: '1.125rem',
   lineHeight: '1.75rem',
   letterSpacing: '0.025rem',
-});
-
-const postBody = css({
-  w: '100%',
-  fontWeight: '300',
-
-  '& h1, h2, h3, h4, h5, h6': {
-    mt: '4rem',
-  },
-
-  '& h1, h2': {
-    mb: '0.5rem',
-    py: '0.5rem',
-    fontWeight: 'bold',
-    borderBottom: '0.0625rem solid #b2ebf24d',
-  },
-
-  '& h1': {
-    fontSize: '1.875rem',
-    lineHeight: '2.25rem',
-  },
-
-  '& h2': {
-    fontSize: '1.5rem',
-    lineHeight: '2rem',
-  },
-
-  '& h3': {
-    _before: {
-      mr: '0.5rem',
-      content: '"#"',
-    },
-  },
-
-  '& h4': {
-    _before: {
-      mr: '0.5rem',
-      content: '"##"',
-    },
-  },
-
-  '& h5': {
-    _before: {
-      mr: '0.5rem',
-      content: '"###"',
-    },
-  },
-
-  '& h3, h4, h5, h6': {
-    fontSize: '1.25rem',
-    lineHeight: '1.75rem',
-    mb: '0.75rem',
-    fontWeight: '400',
-  },
-
-  '& a': {
-    color: '#e879f9',
-    mx: '1px',
-    textDecoration: 'underline',
-  },
-
-  '& figure': {
-    '& img': {
-      w: '100%',
-      mx: 'auto',
-      my: '2rem',
-      borderRadius: '0.5rem',
-    },
-  },
-
-  '& ul': {
-    ml: '1.75rem',
-    mb: '2rem',
-    listStyleType: 'disc',
-
-    '& li': {
-      mx: '0',
-      opacity: '0.7',
-      fontSize: '1rem',
-    },
-  },
-
-  '& hr': {
-    my: '2rem',
-    h: '1px',
-    border: 'none',
-    bg: '#b2ebf24d',
-  },
-
-  '& p': {
-    mb: '2rem',
-    opacity: '0.7',
-    fontSize: '1rem',
-
-    '& code': {
-      p: '0.25rem 0.5rem',
-      mx: '0.25rem',
-      borderRadius: '0.25rem',
-      fontSmoothing: 'antialiased',
-      bg: '#2b3047',
-      color: '#f2f4ffcc',
-    },
-  },
-
-  '& pre': {
-    my: '3rem',
-    p: '1rem',
-    borderRadius: '0.5rem',
-    bg: '#2b3047',
-    color: '#f2f4ffcc',
-
-    '& code': {
-      py: '1rem',
-      fontSize: '1rem',
-    },
-  },
 });
