+++ src/style/style.ts
@@ -1,0 +1,31 @@
+import { css } from '~/styled-system/css';
+
+export const textSm = css({
+  fontSize: '0.875rem',
+  lineHeight: '1.25rem',
+});
+
+export const textLg = css({
+  fontSize: '1.125rem',
+  lineHeight: '1.75rem',
+});
+
+export const textXl = css({
+  fontSize: '1.25rem',
+  lineHeight: '1.75rem',
+});
+
+export const text2xl = css({
+  fontSize: '1.5rem',
+  lineHeight: '2rem',
+});
+
+export const text3xl = css({
+  fontSize: '1.875rem',
+  lineHeight: '2.25rem',
+});
+
+export const text4xl = css({
+  fontSize: '2.25rem',
+  lineHeight: '2.5rem',
+});
